feat: Add password visibility toggle, fix charging history parsing, and improve map UX
- Add password visibility toggle icon to login and complete profile screens
- Fix complete profile password field to accept only numbers
- Fix charging history parsing error by using double.parse instead of int.parse for decimal values
- Remove remaining time display from charger screen
- Add auto-zoom to user location on map first load
- Remove zoom controls (+/-) from map and add locate me button

+++ lib/presentation/history/cubit/history_cubit.dart
@@ -97,15 +97,15 @@ class HistoryCubit extends Cubit<HistoryState> {
         break;
       case 'Highest Energy':
         sortedList.sort((a, b) {
-          final aKwh = int.tryParse(a.kwh ?? '0') ?? 0;
-          final bKwh = int.tryParse(b.kwh ?? '0') ?? 0;
+          final aKwh = double.tryParse(a.kwh ?? '0') ?? 0.0;
+          final bKwh = double.tryParse(b.kwh ?? '0') ?? 0.0;
           return bKwh.compareTo(aKwh);
         });
         break;
       case 'Lowest Energy':
         sortedList.sort((a, b) {
-          final aKwh = int.tryParse(a.kwh ?? '0') ?? 0;
-          final bKwh = int.tryParse(b.kwh ?? '0') ?? 0;
+          final aKwh = double.tryParse(a.kwh ?? '0') ?? 0.0;
+          final bKwh = double.tryParse(b.kwh ?? '0') ?? 0.0;
           return aKwh.compareTo(bKwh);
         });
         break;
